(re)move the comma entries
Original rationale for including them: The commas occur before a dependent
clause and various sources I found online state that no comma should appear
there.

For example:
https://www.grammarly.com/blog/punctuation-capitalization/comma-before-and/

With further research it appeared, however, that this is not in fact a hard
rule; sources hinted at the possibility of the comma breaking up complex
sentences.

Not content to remain uneducated, I spent 14 eurobucks on a copy of the
Penguin Guide to Punctuation. It divides all commas into four classes:
Listing, joining, bracketing, and gapping. Both instances here are neither
listing commas as there are only two items. They are not gapping commas
since they do not stand for an ommission. They are not joining commas
because those separate independent clauses. So if anything, these are
bracketing commas which serve to delimit a "weak interruption", i.e. an
additional thought that is separate from the rest of the sentence but not
so much to need dashes or parentheses.

Indeed, bracketing commas are optional unless the sentence is ambiguous
without them. They can be used to add clarity when needed. However, what
they separate must not be an integral part of the statement. I think the
second case might actually fall foul of that.

"A character may serve up to seven terms of service voluntarily and may
leave after any term " - the parts about the maximum number of terms and
the ability to leave are arguably closely related and should form a single
statement. On its own the second statement is not true, as we are reminded
in the parentheses that directly follow it: A character may not leave if
they roll a 12 on their reinlistment throw. But attaching the statement
"may leave after any term" to the first part makes sense, explaining the
voluntary nature of enlistment. Then it also makes sense to clarify the
exception to that in parentheses.

Out of the above rationale, I'm moving the second case to the "status
uncertain" category. Ultimately it depends on the intent of the author.


More sources:

https://owl.purdue.edu/owl/general_writing/punctuation/commas/extended_rules_for_commas.html
https://english.stackexchange.com/questions/451818/punctuation-of-compound-predicate-where-the-first-item-ends-in-a-list
https://english.stackexchange.com/questions/119839/when-to-use-a-comma-before-and
https://english.stackexchange.com/questions/42811/comma-before-and-or-and-other-coordinating-conjunctions-in-compound-sentenc
https://english.stackexchange.com/questions/30516/should-i-use-a-comma-before-and-or-or
https://www.bbc.co.uk/worldservice/learningenglish/grammar/learnit/learnitv258.shtml
https://www.grammarbook.com/punctuation/commas.asp